오류체크 끝
