+++ checker/checker.dSYM/Contents/Info.plist
