+++ checker/command_1.c
@@ -1,4 +1,4 @@
-#include "push_swap.h"
+#include "checker.h"
 
 void	pa(t_info info)
 {
@@ -10,7 +10,6 @@ void	pa(t_info info)
 	push_head(info.stack_a, node);
 	info.stack_a->size++;
 	info.stack_b->size--;
-	ft_putendl_fd("pa", 1);
 }
 
 void	pb(t_info info)
@@ -23,21 +22,18 @@ void	pb(t_info info)
 	push_head(info.stack_b, node);
 	info.stack_a->size--;
 	info.stack_b->size++;
-	ft_putendl_fd("pb", 1);
 }
 
 void	sa(t_stack *stack_a)
 {
 	if (ft_swap(stack_a) == 0)
 		return ;
-	ft_putendl_fd("sa", 1);
 }
 
 void	sb(t_stack *stack_b)
 {
 	if (ft_swap(stack_b) == 0)
 		return ;
-	ft_putendl_fd("sb", 1);
 }
 
 void	ss(t_info info)
@@ -46,5 +42,4 @@ void	ss(t_info info)
 		return ;
 	if (ft_swap(info.stack_b) == 0)
 		return ;
-	ft_putendl_fd("ss", 1);
 }
