+++ checker/command_2.c
@@ -1,0 +1,27 @@
+#include "checker.h"
+
+void	ra(t_stack *stack_a)
+{
+	t_node	*node;
+
+	node = pop_head(stack_a);
+	if (node == 0)
+		return ;
+	push_tail(stack_a, node);
+}
+
+void	rb(t_stack *stack_b)
+{
+	t_node	*node;
+
+	node = pop_head(stack_b);
+	if (node == 0)
+		return ;
+	push_tail(stack_b, node);
+}
+
+void	rr(t_info info)
+{
+	ra(info.stack_a);
+	rb(info.stack_b);
+}
